Refactored as a package
