+++ inkind/__init__.py
@@ -1,0 +1,3 @@
+from .proposal import *
+from .contribution import *
+from .directory import *
